+++ packages/familie-backend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.16](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@5.0.15...@navikt/familie-backend@5.0.16) (2020-11-18)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+
+
+
+
 ## [5.0.15](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@5.0.14...@navikt/familie-backend@5.0.15) (2020-11-04)