+++ packages/familie-header/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.14](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-header@2.0.13...@navikt/familie-header@2.0.14) (2020-11-18)
+
+**Note:** Version bump only for package @navikt/familie-header
+
+
+
+
+
 ## [2.0.13](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-header@2.0.12...@navikt/familie-header@2.0.13) (2020-11-11)
 
 **Note:** Version bump only for package @navikt/familie-header